eb-flask/requirements.txt
boto==2.40.0
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mysql-connector-repackaged==0.3.1
Werkzeug==0.11.10
